refactor. Packages and classes reordered
