+++ src/main/java/com/bytestream_api2/games/controller/GameController.java
@@ -3,10 +3,10 @@ package com.bytestream_api2.games.controller;
 import com.bytestream_api2.games.entity.GameCategory;
 import com.bytestream_api2.games.entity.GameRating;
 import com.bytestream_api2.games.entity.GameRatingDescriptor;
-import com.bytestream_api2.games.exception.EntityNotFoundException;
-import com.bytestream_api2.games.exception.InvalidEntityRelationsException;
-import com.bytestream_api2.games.exception.MediaUploadFailedException;
-import com.bytestream_api2.games.utilities.BodyFormatter;
+import com.bytestream_api2.games.exception.services.EntityNotFoundException;
+import com.bytestream_api2.games.exception.services.InvalidEntityRelationsException;
+import com.bytestream_api2.games.exception.services.MediaUploadFailedException;
+import com.bytestream_api2.games.utilities.statics.BodyFormatter;
 import com.bytestream_api2.games.validation_groups.Game.OnCreate;
 import com.bytestream_api2.games.validation_groups.Game.OnUpdate;
 import org.springframework.beans.factory.annotation.Autowired;
