+++ src/main/java/com/bytestream_api2/games/converter/GameRatingConverter.java
@@ -1,9 +1,8 @@
 package com.bytestream_api2.games.converter;
 
 import com.bytestream_api2.games.entity.GameRating;
-import com.bytestream_api2.games.interfaces.ImageContentEntity;
+import com.bytestream_api2.games.utilities.interfaces.ImageContentEntity;
 import com.bytestream_api2.games.model.MGameRating;
-import com.bytestream_api2.games.model.MGameRatingEntity;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
