+++ src/main/java/com/bytestream_api2/games/utilities/enums/GameArtType.java
@@ -1,4 +1,4 @@
-package com.bytestream_api2.games.enums;
+package com.bytestream_api2.games.utilities.enums;
 
 import org.jetbrains.annotations.NotNull;
