+++ src/main/java/com/bytestream_api2/games/utilities/statics/BodyFormatter.java
@@ -1,4 +1,4 @@
-package com.bytestream_api2.games.utilities;
+package com.bytestream_api2.games.utilities.statics;
 
 import java.util.Map;
