+++ src/test/java/com/bytestream_api2/games/controller/EntityStaticResourceControllerTest.java
@@ -1,10 +1,10 @@
 package com.bytestream_api2.games.controller;
 
-import com.bytestream_api2.games.exception.EntityNotFoundException;
+import com.bytestream_api2.games.exception.services.EntityNotFoundException;
 import com.bytestream_api2.games.exception.imageresource.ImagePathUnresolvedException;
 import com.bytestream_api2.games.exception.imageresource.UncaughtImageExtensionException;
 import com.bytestream_api2.games.exception.imageresource.UnreadableResourceException;
-import com.bytestream_api2.games.misc.ImageResourcePackage;
+import com.bytestream_api2.games.utilities.classes.ImageResourcePackage;
 import com.bytestream_api2.games.service.GameRatingEntityService;
 import com.bytestream_api2.games.service.GameRatingService;
 import com.bytestream_api2.games.service.GameService;
