+++ src/test/java/com/bytestream_api2/games/service/GameCategoryServiceTest.java
@@ -2,7 +2,7 @@ package com.bytestream_api2.games.service;
 
 import com.bytestream_api2.games.converter.GameCategoryConverter;
 import com.bytestream_api2.games.entity.GameCategory;
-import com.bytestream_api2.games.exception.EntityNotFoundException;
+import com.bytestream_api2.games.exception.services.EntityNotFoundException;
 import com.bytestream_api2.games.mapper.GameCategoryMapper;
 import com.bytestream_api2.games.model.MGameCategory;
 import com.bytestream_api2.games.repository.GameCategoryRepository;
