+++ src/test/java/com/bytestream_api2/games/service/GameRatingDescriptorServiceTest.java
@@ -3,7 +3,7 @@ package com.bytestream_api2.games.service;
 import com.bytestream_api2.games.converter.GameRatingDescriptorConverter;
 import com.bytestream_api2.games.entity.GameRatingDescriptor;
 import com.bytestream_api2.games.entity.GameRatingEntity;
-import com.bytestream_api2.games.exception.EntityNotFoundException;
+import com.bytestream_api2.games.exception.services.EntityNotFoundException;
 import com.bytestream_api2.games.mapper.GameRatingDescriptorMapper;
 import com.bytestream_api2.games.model.MGameRatingDescriptor;
 import com.bytestream_api2.games.repository.GameRatingDescriptorRepository;
