+++ src/test/java/com/bytestream_api2/games/service/GameServiceTest.java
@@ -5,23 +5,22 @@ import com.bytestream_api2.games.entity.Game;
 import com.bytestream_api2.games.entity.GameCategory;
 import com.bytestream_api2.games.entity.GameRating;
 import com.bytestream_api2.games.entity.GameRatingDescriptor;
-import com.bytestream_api2.games.exception.EntityNotFoundException;
-import com.bytestream_api2.games.exception.InvalidEntityRelationsException;
-import com.bytestream_api2.games.exception.MediaUploadFailedException;
+import com.bytestream_api2.games.exception.services.EntityNotFoundException;
+import com.bytestream_api2.games.exception.services.InvalidEntityRelationsException;
+import com.bytestream_api2.games.exception.services.MediaUploadFailedException;
 import com.bytestream_api2.games.exception.imageresource.ImagePathUnresolvedException;
 import com.bytestream_api2.games.exception.imageresource.UncaughtImageExtensionException;
 import com.bytestream_api2.games.exception.imageresource.UnreadableResourceException;
-import com.bytestream_api2.games.interfaces.ImageContentEntity;
+import com.bytestream_api2.games.utilities.interfaces.ImageContentEntity;
 import com.bytestream_api2.games.mapper.GameMapper;
-import com.bytestream_api2.games.misc.ImageResourcePackage;
+import com.bytestream_api2.games.utilities.classes.ImageResourcePackage;
 import com.bytestream_api2.games.model.MGame;
-import com.bytestream_api2.games.model.MGameRating;
 import com.bytestream_api2.games.repository.GameCategoryRepository;
 import com.bytestream_api2.games.repository.GameRatingDescriptorRepository;
 import com.bytestream_api2.games.repository.GameRatingRepository;
 import com.bytestream_api2.games.repository.GameRepository;
-import com.bytestream_api2.games.utilities.EntityResolver;
-import com.bytestream_api2.games.utilities.ImageResourceManager;
+import com.bytestream_api2.games.utilities.statics.EntityResolver;
+import com.bytestream_api2.games.utilities.statics.ImageResourceManager;
 import jakarta.servlet.ServletContext;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
